# ground-control/swagger.yml
basePath: /
consumes:
    - application/json
definitions:
...
    APIEmptyObject:
        title: APIEmptyObject is an empty JSON object response.
        type: object
        x-go-package: github.com/container-registry/harbor-satellite/ground-control/internal/server
    APIGroup:
        allOf:
            - properties:
                CreatedAt:
                    format: date-time
                    type: string
                GroupName:
                    type: string
                ID:
                    format: int32
                    type: integer
                Projects:
                    items:
                        type: string
                    type: array
                RegistryUrl:
                    type: string
                UpdatedAt:
                    format: date-time
                    type: string
              type: object
        title: APIGroup describes a satellite group row.
        x-go-package: github.com/container-registry/harbor-satellite/ground-control/internal/server
...
    APIMessageResponse:
        properties:
            message:
                type: string
                x-go-name: Message
        title: APIMessageResponse contains a human-readable command result.
        type: object
        x-go-package: github.com/container-registry/harbor-satellite/ground-control/internal/server
...
    APISatellite:
        allOf:
            - properties:
                CreatedAt:
                    format: date-time
                    type: string
                HeartbeatInterval:
                    $ref: '#/definitions/NullString'
                ID:
                    format: int32
                    type: integer
                LastSeen:
                    $ref: '#/definitions/NullTime'
                Name:
                    type: string
                UpdatedAt:
                    format: date-time
                    type: string
              type: object
        title: APISatellite describes a satellite row.
        x-go-package: github.com/container-registry/harbor-satellite/ground-control/internal/server
...
    AppError:
        properties:
            code:
                format: int64
                type: integer
                x-go-name: Code
            message:
                type: string
                x-go-name: Message
        title: AppError is a structured JSON error response.
        type: object
        x-go-package: github.com/container-registry/harbor-satellite/ground-control/internal/server
...
    NullString:
        description: |-
            var s NullString
            err := db.QueryRow("SELECT name FROM foo WHERE id=?", id).Scan(&s)
            ...
            if s.Valid {
            use s.String
            } else {
            NULL value
            }
        properties:
            String:
                type: string
            Valid:
                type: boolean
        title: |-
            NullString represents a string that may be null.
            NullString implements the [Scanner] interface so
            it can be used as a scan destination:
        type: object
        x-go-package: database/sql
    NullTime:
        description: |-
            NullTime implements the [Scanner] interface so
            it can be used as a scan destination, similar to [NullString].
        properties:
            Time:
                format: date-time
                type: string
            Valid:
                type: boolean
        title: NullTime represents a [time.Time] that may be null.
        type: object
        x-go-package: database/sql
...
    SatelliteGroupParams:
        properties:
            group:
                type: string
                x-go-name: Group
            satellite:
                type: string
                x-go-name: Satellite
        title: SatelliteGroupParams links or unlinks a satellite and a group.
        type: object
        x-go-package: github.com/container-registry/harbor-satellite/ground-control/internal/server
...
info:
    description: |
        Ground Control manages Harbor Satellite registration, configuration,
        group membership, user administration, status reporting, and SPIRE/SPIFFE
        bootstrap flows.
    title: Ground Control API
    version: 0.1.0
paths:
...
    /api/groups:
        get:
            operationId: listGroups
            responses:
                "200":
                    description: Groups returned.
                    schema:
                        items:
                            $ref: '#/definitions/APIGroup'
                        type: array
                "401":
                    description: Authorization token is missing or invalid.
                    schema:
                        $ref: '#/definitions/AppError'
                "500":
                    description: Groups could not be listed.
                    schema:
                        $ref: '#/definitions/AppError'
            security:
                - bearerAuth: []
            summary: Lists groups.
            tags:
                - groups
    /api/groups/{group}:
        delete:
            operationId: deleteGroup
            parameters:
                - description: Group name.
                  in: path
                  name: group
                  required: true
                  type: string
                  x-go-name: Group
            responses:
                "200":
                    description: Group deleted.
                    schema:
                        $ref: '#/definitions/APIEmptyObject'
                "401":
                    description: Authorization token is missing or invalid.
                    schema:
                        $ref: '#/definitions/AppError'
                "403":
                    description: System administrator role is required.
                    schema:
                        $ref: '#/definitions/AppError'
                "404":
                    description: Group was not found.
                    schema:
                        $ref: '#/definitions/AppError'
                "500":
                    description: Group could not be deleted.
                    schema:
                        $ref: '#/definitions/AppError'
            security:
                - bearerAuth: []
            summary: Deletes a group.
            tags:
                - groups
        get:
            operationId: getGroup
            parameters:
                - description: Group name.
                  in: path
                  name: group
                  required: true
                  type: string
                  x-go-name: Group
            responses:
                "200":
                    description: Group returned.
                    schema:
                        $ref: '#/definitions/APIGroup'
                "401":
                    description: Authorization token is missing or invalid.
                    schema:
                        $ref: '#/definitions/AppError'
                "404":
                    description: Group was not found.
                    schema:
                        $ref: '#/definitions/AppError'
                "500":
                    description: Group could not be loaded.
                    schema:
                        $ref: '#/definitions/AppError'
            security:
                - bearerAuth: []
            summary: Gets a group by name.
            tags:
                - groups
...
    /api/groups/satellite:
        delete:
            operationId: removeSatelliteFromGroup
            parameters:
                - description: Satellite and group membership payload.
                  in: body
                  name: Body
                  required: true
                  schema:
                    $ref: '#/definitions/SatelliteGroupParams'
            responses:
                "200":
                    description: Satellite removed from the group.
                    schema:
                        $ref: '#/definitions/APIEmptyObject'
                "400":
                    description: Satellite or group is invalid or not found.
                    schema:
                        $ref: '#/definitions/AppError'
                "401":
                    description: Authorization token is missing or invalid.
                    schema:
                        $ref: '#/definitions/AppError'
                "500":
                    description: Satellite could not be removed from the group.
                    schema:
                        $ref: '#/definitions/AppError'
            security:
                - bearerAuth: []
            summary: Removes a satellite from a group.
            tags:
                - groups
        post:
            operationId: addSatelliteToGroup
            parameters:
                - description: Satellite and group membership payload.
                  in: body
                  name: Body
                  required: true
                  schema:
                    $ref: '#/definitions/SatelliteGroupParams'
            responses:
                "200":
                    description: Satellite was added to the group or was already a member.
                    schema:
                        $ref: '#/definitions/APIMessageResponse'
                "400":
                    description: Satellite or group is invalid or not found.
                    schema:
                        $ref: '#/definitions/AppError'
                "401":
                    description: Authorization token is missing or invalid.
                    schema:
                        $ref: '#/definitions/AppError'
                "500":
                    description: Satellite could not be added to the group.
                    schema:
                        $ref: '#/definitions/AppError'
            security:
                - bearerAuth: []
            summary: Adds a satellite to a group.
            tags:
                - groups
...
    /api/satellites/{satellite}:
        delete:
            operationId: deleteSatellite
            parameters:
                - description: Satellite name.
                  in: path
                  name: satellite
                  required: true
                  type: string
                  x-go-name: Satellite
            responses:
                "200":
                    description: Satellite deleted.
                    schema:
                        $ref: '#/definitions/APIEmptyObject'
                "400":
                    description: Satellite does not exist or cannot be deleted.
                    schema:
                        $ref: '#/definitions/AppError'
                "401":
                    description: Authorization token is missing or invalid.
                    schema:
                        $ref: '#/definitions/AppError'
                "500":
                    description: Satellite could not be deleted.
                    schema:
                        $ref: '#/definitions/AppError'
            security:
                - bearerAuth: []
            summary: Deletes a satellite.
            tags:
                - satellites
        get:
            operationId: getSatellite
            parameters:
                - description: Satellite name.
                  in: path
                  name: satellite
                  required: true
                  type: string
                  x-go-name: Satellite
            responses:
                "200":
                    description: Satellite returned.
                    schema:
                        $ref: '#/definitions/APISatellite'
                "401":
                    description: Authorization token is missing or invalid.
                    schema:
                        $ref: '#/definitions/AppError'
                "500":
                    description: Satellite could not be loaded.
                    schema:
                        $ref: '#/definitions/AppError'
            security:
                - bearerAuth: []
            summary: Gets a satellite by name.
            tags:
                - satellites
...
produces:
    - application/json
...
schemes:
    - https
securityDefinitions:
    bearerAuth:
        description: 'Bearer token authentication. Use the format: Bearer <token>.'
        in: header
        name: Authorization
        type: apiKey
swagger: "2.0"